Fix memory leaks
The worst offender was the optional src3 matrix passed into `gemm`.
It was being created in the function call so I missed it when making
sure all OpenCV.js objects were deleted. Since `applyHomography` is
called so often in the rendering the tool path on top of the camera
view, a memory leak there will have a magnified impact.